requirements.txt
numpy
pillow
tqdm
scikit-learn
faiss-cpu
torch
torchvision
transformers
